Implement more correct init logic, caching 'all' calls, and other minor cleanups

+++ caching_store_wrapper.js
@@ -6,52 +6,77 @@ function cacheKey(kind, key) {
   return kind.namespace + ":" + key;
 }
 
+function allCacheKey(kind) {
+  return "$all:" + kind.namespace;
+}
+
+var initializedKey = "$checkedInit";
+
+
 function CachingStoreWrapper(underlyingStore, ttl) {
   var cache = ttl ? new NodeCache({ stdTTL: ttl }) : null;
   var queue = new UpdateQueue();
-  var initialized;
+  var initialized = false;
 
   this.init = function(allData, cb) {
-    initialized = undefined;
-    queue.enqueue(function(cb) {
-      cache.flushAll();
+    if (cache) {
+      queue.enqueue(function(cb) {
+        cache.del(initializedKey);
+        cache.flushAll();
 
-      // populate cache with initial data
-      for (var kindNamespace in allData) {
-        if (Object.hasOwnProperty.call(allData, kindNamespace)) {
-          var kind = dataKind[kindNamespace];
-          var items = allData[kindNamespace];
-          for (var key in items) {
-            cache.set(kind, key, items[key]);
+        // populate cache with initial data
+        for (var kindNamespace in allData) {
+          if (Object.hasOwnProperty.call(allData, kindNamespace)) {
+            var kind = dataKind[kindNamespace];
+            var items = allData[kindNamespace];
+            cache.set(allCacheKey(kind), items);
+            for (var key in items) {
+              cache.set(cacheKey(kind, key), items[key]);
+            }
           }
         }
-      }
 
-      underlyingStore.init(allData, cb);
-    }, [], cb);
+        underlyingStore.init(allData, cb);
+      }, [], cb);
+    }
   };
 
   this.initialized = function(cb) {
-    if (initialized === undefined) {
+    if (initialized) {
+      cb(initialized);
+    } else if (cache && cache.get(initializedKey)) {
+      return false;
+    } else {
       underlyingStore.initialized(function(inited) {
         initialized = inited;
+        if (!initialized) {
+          cache && cache.set(initializedKey, true);
+        }
         cb(initialized);
       });
-    } else {
-      // if we already have a cached value for initialized, don't bother trying
-      // to call the underlying store until someone calls init, since it won't
-      // change.
-      cb(initialized);
     }
   };
 
   this.all = function(kind, cb) {
-    underlyingStore.all(kind, cb);
+    var items = cache && cache.get(allCacheKey(kind));
+    if (items) {
+      cb(items);
+      return;
+    }
+
+    underlyingStore.all(kind, function(items) {
+      cache.set(allCacheKey(kind), items)
+      cb(items);
+    });
   };
 
   this.get = function(kind, key, cb) {
-    item = cache && cache.get(cacheKey(kind, key));
-    if (item && !item.deleted) {
+    var item = cache && cache.get(cacheKey(kind, key));
+    if (item) {
+      if (item.deleted) {
+        cb(null);
+        return;
+      }
       cb(item);
       return;
     }
@@ -64,9 +89,10 @@ function CachingStoreWrapper(underlyingStore, ttl) {
 
   this.upsert = function(kind, newItem, cb) {
     queue.enqueue(function (cb) {
+      flushAllCaches();
       underlyingStore.upsertInternal(kind, newItem, function(err, attemptedWrite) {
         if (attemptedWrite && !err) {
-          var red = cache && cache.set(cacheKey(kind, newItem.key), newItem);
+          cache && cache.set(cacheKey(kind, newItem.key), newItem);
         }
         cb();
       });
@@ -81,6 +107,15 @@ function CachingStoreWrapper(underlyingStore, ttl) {
     cache.close();
     underlyingStore.close();
   };
+
+  function flushAllCaches() {
+    if (!cache) {
+      return;
+    }
+    for (var kindNamespace in dataKind) {
+      cache.del(allCacheKey(dataKind[kindNamespace]));
+    }
+  }
 }
 
 module.exports = CachingStoreWrapper;
